Add ProductStock policy

+++ app/Http/Controllers/ProductStockController.php
@@ -7,11 +7,15 @@ use App\Http\Requests\StoreProductStockRequest;
 use App\Http\Requests\UpdateProductStockRequest;
 use App\Http\Resources\ProductStockResource;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Gate;
 
 class ProductStockController extends Controller
 {
     public function index()
     {
+        // Gate
+        Gate::authorize('viewAny', ProductStock::class);
+
         $data = ProductStock::with('product', 'amountType')->get();
 
         return response()->json([
@@ -22,6 +26,9 @@ class ProductStockController extends Controller
 
     public function store(StoreProductStockRequest $request): JsonResponse
     {
+        // Gate
+        Gate::authorize('creeate', ProductStock::class);
+
         $newStock = ProductStock::create($request->validated());
 
         return response()->json([
@@ -33,6 +40,9 @@ class ProductStockController extends Controller
 
     public function show(ProductStock $productStock)
     {
+        // Gate
+        Gate::authorize('view', ProductStock::class);
+
         return response()->json([
             "data" => ProductStockResource::collection($productStock),
         ]);
@@ -41,6 +51,9 @@ class ProductStockController extends Controller
 
     public function update(UpdateProductStockRequest $request, ProductStock $productStock)
     {
+        // Gate
+        Gate::authorize('update', ProductStock::class);
+
         // Check if Name already exists
         if ($request->validated('name')) {
             $nameExist = ProductStock::where('name', $request->validated('name'))->where('id', '<>', $productStock->id)->exists();
@@ -58,6 +71,9 @@ class ProductStockController extends Controller
 
     public function destroy(ProductStock $productStock): JsonResponse
     {
+        // Gate
+        Gate::authorize('delete', ProductStock::class);
+
         $productStock->delete();
 
         return response()->json([
